requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
altgraph==0.17.2
asttokens==2.0.5
astunparse==1.6.3
async-timeout==4.0.2
attrs==21.4.0
backcall==0.2.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
cycler==0.11.0
decorator==5.1.1
executing==0.8.3
fastjsonschema==2.15.3
flatbuffers==2.0
fonttools==4.33.3
frozenlist==1.3.0
gast==0.5.3
google-auth==2.6.5
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.2
grpcio==1.44.0
h5py==3.6.0
idna==3.3
imageio==2.19.0
importlib-metadata==4.11.3
importlib-resources==5.7.1
ipython==8.3.0
jedi==0.18.1
Jinja2==3.1.1
joblib==1.1.0
jsonschema==4.4.0
jupyter-core==4.10.0
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
libclang==13.0.0
Markdown==3.3.6
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
multidict==6.0.2
nbformat==5.4.0
networkx==2.8
numpy==1.22.3
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.2
parso==0.8.3
pdfkit==1.0.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
plotly==5.7.0
prompt-toolkit==3.0.29
protobuf==3.20.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydicom==2.3.0
Pygments==2.12.0
pyinstaller==4.10
pyinstaller-hooks-contrib==2022.3
pyparsing==3.0.8
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.10.1
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.1
PyWavelets==1.3.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
SQLAlchemy==1.4.36
stack-data==0.2.0
tenacity==8.0.1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.1.0
tifffile==2022.5.4
traitlets==5.1.1
typing-extensions==4.1.1
urllib3==1.26.9
vtk==9.1.0
wcwidth==0.2.5
Werkzeug==2.1.1
wrapt==1.14.0
wslink==1.6.1
yarl==1.7.2
zipfile36==0.1.3
zipp==3.8.0
